SeeSign

Documentation for SeeSign.

SeeSign.StepArray

Base.iterate

SeeSign.accept

SeeSign.changed

SeeSign.changed

SeeSign.deal_with_changes

SeeSign.isadjacent

SeeSign.move_generate_event

SeeSign.previous_value

SeeSign.track_assignments

SeeSign.StepArray — Type

StepArray(arr)

You have an array and want to compare the current value with the previous value. This creates a double of the array so that memory use is optimized.

Base.iterate — Function

iterate(arr::StepArray, [state])

Enables iteration over a StepArray, allowing it to be used in for loops and with other iteration functions like map, filter, etc.

SeeSign.accept — Method

The arrays in a PhysicalState record that they have been modified. This function erases the record of modifications.

SeeSign.changed — Method

Return a list of changed places in the physical state. A place for this state is a tuple of a symbol and the Cartesian index. The symbol is the name of the array within the PhysicalState.

SeeSign.changed — Method

changed(arr::StepArray)

SeeSign.deal_with_changes — Method

deal_with_changes(sim::SimulationFSM)

An event changed the state. This function modifies events to respond to changes in state.

SeeSign.isadjacent — Method

Whether two individuals are adjacent.

SeeSign.move_generate_event — Method

A dynamic generator of events. This looks at a changed place in the physical state and generates clock keys for events that could depend on that place.

@when agentat(loc) && available(loc+direction) => (:move, loc, direction)

SeeSign.previous_value — Method

previous_value(arr::StepArray, i...)

Get the previous value at index i before it was changed. Returns the current value if the element hasn't been changed.

SeeSign.track_assignments — Method

track_assignments(f::Function)

This function takes a function f that returns nothing and creates a new function that returns a list of the indices of the elements that are assigned to in f.

It uses code introspection to analyze the function for array assignments without having to execute it with a proxy object.