SeeSign

Documentation for SeeSign.

SeeSign.DependencyNetwork

SeeSign.StepArray

SeeSign.TrackedVector

Base.iterate

Base.show

SeeSign.accept

SeeSign.allowed_infects

SeeSign.allowed_moves

SeeSign.ascii_to_array

SeeSign.change

SeeSign.changed

SeeSign.changed

SeeSign.changed

SeeSign.check_events

SeeSign.deal_with_changes

SeeSign.enable

SeeSign.gotten

SeeSign.isconsistent

SeeSign.isconsistent

SeeSign.move_agent

SeeSign.over_tracked_physical_state

SeeSign.previous_value

SeeSign.reset_gotten!

SeeSign.reset_tracking!

SeeSign.resetread

SeeSign.sick_movement

SeeSign.toencroach_generate_event

SeeSign.tomove_generate_event

SeeSign.tosickfriend_generate_event

SeeSign.tospace_generate_event

SeeSign.track_assignments

SeeSign.wasread

SeeSign.@tracked_struct

SeeSign.DependencyNetwork — Type

This is a multi-graph from places to events. It is a multi-graph because there can be two kinds of edges from the same place to the same event: an enabling edge and a rate edge. The graph is mutable, too, so you can add and remove events which adds and removes edges from the graph.

SeeSign.StepArray — Type

StepArray(arr)

You have an array and want to compare the current value with the previous value. This creates a double of the array so that memory use is optimized.

SeeSign.TrackedVector — Type

TrackedVector{T}

A vector that tracks access and changes to its elements.

Base.iterate — Function

iterate(arr::StepArray, [state])

Enables iteration over a StepArray, allowing it to be used in for loops and with other iteration functions like map, filter, etc.

Base.show — Method

Pretty print the board state.

SeeSign.accept — Method

The arrays in a PhysicalState record that they have been modified. This function erases the record of modifications.

SeeSign.allowed_infects — Method

For debugging, look at every allowed infection.

SeeSign.allowed_moves — Method

For debugging, list every allowed movement transition.

SeeSign.ascii_to_array — Method

For testing, convert an ASCII image of a board into a 2D array of integers.

SeeSign.change — Method

change(obj)

Returns the set of fields that have been modified. Alias for changed().

SeeSign.changed — Method

Return a list of changed places in the physical state. A place for this state is a tuple of a symbol and the Cartesian index. The symbol is the name of the array within the PhysicalState.

SeeSign.changed — Method

changed(arr::StepArray)

SeeSign.changed — Method

changed(obj)

Returns the set of fields that have been modified.

SeeSign.check_events — Method

More debugging, check that all events are correct.

SeeSign.deal_with_changes — Method

deal_with_changes(sim::SimulationFSM)

An event changed the state. This function modifies events to respond to changes in state.

SeeSign.enable — Method

This function decides the rate of the transition, but whether the transition is enabled was already decided by the @condition in the macro. That same @condition will be used to disable the transition.

SeeSign.gotten — Method

gotten(obj)

Returns the set of fields that have been accessed.

SeeSign.isconsistent — Method

Used for debugging.

SeeSign.isconsistent — Method

Double-check the board state.

SeeSign.move_agent — Method

Because the array of board squares is 1D but the board is 2D, there is translation when you move an agent.

SeeSign.over_tracked_physical_state — Method

Iterate over all tracked vectors in the physical state.

SeeSign.previous_value — Method

previous_value(arr::StepArray, i...)

Get the previous value at index i before it was changed. Returns the current value if the element hasn't been changed.

SeeSign.reset_gotten! — Method

reset_gotten!(obj)

Reset the tracking of accessed fields.

SeeSign.reset_tracking! — Method

reset_tracking!(obj)

Reset all tracking information.

SeeSign.resetread — Method

Return a list of changed places in the physical state. A place for this state is a tuple of a symbol and the Cartesian index. The symbol is the name of the array within the PhysicalState.

SeeSign.sick_movement — Method

I found writing the @condition macro to be complicated, so I wrote a helper function. The trick with the @condition macro is that it will run inside a begin-end block, so there is no way to short-circuit the evalutation. You can write code like that, but it gets complicated. This macro uses short-cicuiting to make it easier.

SeeSign.toencroach_generate_event — Method

Let's add a process for susceptible-infected. There are two cases to handle that have to do with movement.

Susceptible moves next to infected.

Infected moves next to susceptible.

SeeSign.tomove_generate_event — Method

A dynamic generator of events. This looks at a changed place in the physical state and generates clock keys for events that could depend on that place.

SeeSign.tosickfriend_generate_event — Method

This is the case where a neighbor didn't move but became infected in place and can therefore now infect a neighbor.

SeeSign.tospace_generate_event — Method

This is the case where a neighbor couldn't move because it was blocked but now the space is free and it can move.

SeeSign.track_assignments — Method

track_assignments(f::Function)

This function takes a function f that returns nothing and creates a new function that returns a list of the indices of the elements that are assigned to in f.

It uses code introspection to analyze the function for array assignments without having to execute it with a proxy object.

SeeSign.wasread — Method

Return a list of changed places in the physical state. A place for this state is a tuple of a symbol and the Cartesian index. The symbol is the name of the array within the PhysicalState.

SeeSign.@tracked_struct — Macro

@tracked_struct Name begin
    field1::Type1
    field2::Type2
    # ...
end

Creates a struct that tracks when its fields are accessed or modified.